New course page layout
Changed the layout of the course page to have the nesting of the tabs
sit below each offering (which makes more sense). Also reorganised the
tabs while at it. New layout looks more cluttered, which can hopefully
be fixed (but in fairness there is now also more information being
displayed).